training/models/co2/effcc_distilled_main/requirements.txt
torch
pandas
numpy
pillow
torchvision
